Add function for computing the row pointers of a matrix
Reuse this function for sorting matrices.

Also, add a function for counting the number of stored diagonal
entries in each row of a matrix. This can be used to compute the
degree of each vertex whenever a sparse matrix with a symmetric
sparsity pattern is viewed as the adjacency matrix of an undirected
graph.